requirements.txt
cycler==0.10.0
numpy==1.16.1
scipy==1.2.1
setuptools==40.8.0
matplotlib==3.0.2
ruamel.base==1.0.0
